Fixed Issues with FrontEnd Connection
Fixed problem with Front-End Connection, webconfig does not work well with Spring Boot Security when we added the Security Implementaion had to create a new config that works well and able to connect to Front End. Commented out all of WebConfig code. I think we cam delete it but to be safe i kept